Add to common_queries.md
I mostly want to have a list of queries that I know I'll need, and then
I can either try to translate them into Jetquery SQL commands, or use the
raw SQL functionality of Jetquery. Need to look at the code for date
checking, it's expecting both dates to be set if either one is set.